# api/openapi/marketplace.v1.yaml
openapi: 3.1.0
info:
  title: Marketplace API
  version: 1.0.0
  description: Публичный REST API маркетплейса. Контракт является источником правды (contract-first).
servers:
  - url: https://api.marketplace.kz
  - url: http://localhost:8080
...
paths:
...
  /api/v1/catalog/categories/{id}/facets:
    get:
      tags: [search]
      operationId: getCategoryFacets
      description: Фасеты по опубликованным товарам поддерева категории с учётом текущих фильтров.
      parameters:
        - $ref: "#/components/parameters/PathID"
        - { name: q, in: query, schema: { type: string } }
        - { name: filters, in: query, schema: { type: string }, description: "Список `код:значение`, значения через `|`, пары через запятую" }
        - { name: price_min, in: query, schema: { type: integer, format: int64 } }
        - { name: price_max, in: query, schema: { type: integer, format: int64 } }
        - { name: seller, in: query, schema: { type: string, format: uuid } }
        - { name: condition, in: query, schema: { type: string, enum: [new, used, refurbished] } }
        - { name: in_stock, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Фасеты категории
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Facets" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/search/products:
    get:
      tags: [search]
      operationId: searchProducts
      description: Полнотекстовый поиск с морфологией; при нулевой выдаче выполняется одна попытка исправления опечаток.
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string, format: uuid } }
        - { name: filters, in: query, schema: { type: string } }
        - { name: price_min, in: query, schema: { type: integer, format: int64 } }
        - { name: price_max, in: query, schema: { type: integer, format: int64 } }
        - { name: seller, in: query, schema: { type: string, format: uuid } }
        - { name: condition, in: query, schema: { type: string, enum: [new, used, refurbished] } }
        - { name: in_stock, in: query, schema: { type: boolean } }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, price_asc, price_desc, new, popular] } }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Результаты поиска
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/admin/search/reindex:
    post:
      tags: [search-admin]
      operationId: requestReindex
      description: "Полная переиндексация без простоя: сборка в неактивной версии таблицы и переключение view."
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Задание поставлено в очередь
          content:
            application/json:
              schema:
                type: object
                required: [job_id]
                properties:
                  job_id: { type: string, format: uuid }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/admin/search/reindex/{id}:
    get:
      tags: [search-admin]
      operationId: getReindexJob
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Состояние задания переиндексации
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReindexJob" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    BadRequest:
      description: Некорректный запрос
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Forbidden:
      description: Операция запрещена
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Ресурс не найден
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: Конфликт состояния
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        trace_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
    PageInfo:
      type: object
      required: [has_more]
      properties:
        next_cursor: { type: string }
        has_more: { type: boolean }
...
    SearchHit:
      type: object
      required: [product_id, seller_id, category_id, title, min_price, currency, offers, sellers, in_stock, published_at]
      properties:
        product_id: { type: string, format: uuid }
        seller_id: { type: string, format: uuid }
        category_id: { type: string, format: uuid }
        title: { type: string }
        brand: { type: string }
        cover_key: { type: string }
        min_price: { type: integer, format: int64 }
        currency: { type: string }
        offers: { type: integer }
        sellers: { type: integer }
        in_stock: { type: boolean }
        published_at: { type: string, format: date-time }
        rank: { type: number, format: double }
    SearchResult:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/SearchHit" }
        page: { $ref: "#/components/schemas/PageInfo" }
        corrected_query: { type: string, description: Заполняется, если исходный запрос не дал результатов и был исправлен }
    Facets:
      type: object
      required: [total, price_min, price_max, attributes, numbers, conditions]
      properties:
        total: { type: integer }
        price_min: { type: integer, format: int64 }
        price_max: { type: integer, format: int64 }
        attributes:
          type: array
          items:
            type: object
            required: [code, values]
            properties:
              code: { type: string }
              values:
                type: array
                items:
                  type: object
                  required: [value, count]
                  properties:
                    value: { type: string }
                    count: { type: integer }
        numbers:
          type: array
          items:
            type: object
            required: [code, min, max]
            properties:
              code: { type: string }
              min: { type: number }
              max: { type: number }
        conditions:
          type: array
          items:
            type: object
            required: [value, count]
            properties:
              value: { type: string }
              count: { type: integer }
    ReindexJob:
      type: object
      required: [id, status, target, processed, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [pending, running, completed, failed] }
        target: { type: string, enum: [documents_a, documents_b] }
        processed: { type: integer }
        failure_reason: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
...